random_monge_matrix: simplify matrix generation
We now scale the number of upper-right and lower-left Monge matrices
by the dimensions of the final matrix. This seems to generate "nice"
matrices with a wide spectrum of row-minima.

That is, the minima progress nicely from left to right as you go down
the matrix and are not all clustered in a single column.