recursive_inner: optimize row/column minima
Before we passed an axis argument down the recursive calls and used
that to determine if we were finding row or column minima. We now use
an argument of a generic type instead.

Using a generic argument allows monomorphization to kick in: the
compiler will generate different code for each type of argument. The
types we use are different functions that implement Fn() -> Direction.
Each function will get its own version of the code in recursive_inner
and this means we avoid runtime branching.